\name{RcmdrPlugin.doex-internal}
\title{Internal RcmdrPlugin.doex objects}
\alias{ranblockAnova}
\alias{LatinAnova}
\alias{bartlett}
\alias{GreekAnova}
\alias{factorialAnova}
\alias{twokfactorialAnova}







\description{Internal RcmdrPlugin.doex objects.}
\details{These are not to be called by the user.}
\keyword{internal}
