\name{RcmdrPlugin.doex-package}
\alias{RcmdrPlugin.doex-package}
\alias{RcmdrPlugin.doex}
\docType{package}
\title{
Install the DOEX  Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" for teaching design of experiments
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.doex\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2010-03-28\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Erin Hodgess <hodgesse@uhd.edu>

Maintainer: Erin Hodgess <hodgesse@uhd.edu>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

